requirements.txt
dotenv==0.9.9
fastapi==0.115.12
requests==2.32.3
PyPDF2==3.0.1
python-docx==0.2.4